requirements.txt
virtualenv
langchain
langchain-community
langchain-core
langchain-openai
python-dotenv
beautifulsoup4
faiss-cpu
langchainhub
# "langserve[all]"
wikipedia
elasticsearch
langchain-elasticsearch
llama-index